requirements.txt
Pillow==9.0.1
pygame==2.1.2
PyTMX==3.31
pytweening==1.0.4
